+++ SourceCode/Services/Implementations/PersonService.cs
@@ -3,6 +3,7 @@ using ModulesRegistry.Data;
 using ModulesRegistry.Services.Extensions;
 using ModulesRegistry.Services.Resources;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -32,7 +33,6 @@ namespace ModulesRegistry.Services.Implementations
             using var dbContext = Factory.CreateDbContext();
             return await dbContext.People
                 .Where(p => p.CountryId == inCountryId)
-                .Include(p => p.User)
                 .OrderBy(p => p.FirstName).ThenBy(p => p.LastName).ThenBy(p => p.CityName)
                 .ToListAsync();
         }